Add solution to '02-flex-header' exercise
Add display flex, justify content, align items, and 10px padding to the header element.
This will center the header and the elements around it, as well as give distance
between the container box and its elements.

Add margin auto and to the logo class to push the lists on each side
of the header to the edges of the container box.

Adding display flex property to unordered lists will make the list items
display in a row instead of a column. Setting margin and padding to 0
makes sure there is no unnecessary space between the list items.