Add flexbox styling to the top of webpage
To achieve the standard look, make two div containers, each holding
items that need to be on the left and on the right, respectively.
Add unique class to each container. Both containers will share the same
properties. Making each container a flexbox and adding spacing and
aligning items about secondary axis to the center allows us to achieve
the desires look.

Additionally, removing padding and margin from default ul is also
necessary. Lastly, we need to add the gap property to ul element to
achieve the 16 px gap between the logo and the first link.